Step the Bonsai tape noise LCG in unsigned arithmetic
super_simple_noise is a linear congruential generator:

    return last * 1103515245 + 12345;

Both operands are int32_t, so the multiply overflows for all but a
handful of inputs, and signed overflow is undefined. UBSan reports it
from Surge XT's [fx] suite:

  Bonsai.h:725:63: runtime error: signed integer overflow:
  1103527590 * 1103515245 cannot be represented in type 'int'
    #2 sst::effects::bonsai::Bonsai<..>::tape_noise(..) Bonsai.h:1449
    #3 sst::effects::bonsai::Bonsai<..>::processBlock(float*, float*) Bonsai.h:1576
    #4 surge::sstfx::SurgeSSTFXBase<..>::process(float*, float*)

An LCG is defined in terms of modular arithmetic, so doing the step as
uint32_t - where the wrap is specified rather than undefined - is what
the algorithm wants, and the result converts back unchanged.

The sequence is preserved. Comparing both forms over a million steps
per seed at -O0 through -O3, the outputs are identical for seeds 0, 1,
-1, INT32_MAX and 123456789. They differ only when the state is exactly
INT32_MIN, where the current code is undefined and the optimiser folds
the multiply away, returning last + 12345 instead of stepping.

Surge XT's [fx] suite is unchanged at 3202892 assertions across 9 cases
and the report goes away.

+++ include/sst/effects/Bonsai.h
@@ -722,7 +722,13 @@ template <size_t blockSize> inline void blockabs(float *src)
     }
 }
 
-inline int32_t super_simple_noise(int32_t last) { return last * 1103515245 + 12345; }
+// The multiply overflows int32_t for all but a handful of inputs, and signed overflow
+// is undefined. An LCG wants modular arithmetic anyway, so do the step unsigned, where
+// the wrap is defined, and convert the result back.
+inline int32_t super_simple_noise(int32_t last)
+{
+    return (int32_t)((uint32_t)last * 1103515245u + 12345u);
+}
 template <size_t blockSize> inline void noise(float &last, float *__restrict dst)
 {
     int32_t temp alignas(16)[blockSize] = {};
